+++ examples/siro_sandbox/app_states/app_state_fetch.py
@@ -10,7 +10,10 @@ from app_states.app_state_abc import AppState
 from camera_helper import CameraHelper
 from controllers.gui_controller import GuiHumanoidController
 from gui_navigation_helper import GuiNavigationHelper
-from hablab_utils import get_agent_art_obj_transform, get_grasped_objects_idxs
+from utils.hablab_utils import (
+    get_agent_art_obj_transform,
+    get_grasped_objects_idxs,
+)
 
 from habitat.gui.gui_input import GuiInput
 from habitat.gui.text_drawer import TextOnScreenAlignment
