+++ examples/siro_sandbox/app_states/app_state_socialnav.py
@@ -12,8 +12,8 @@ from app_states.app_state_abc import AppState
 from camera_helper import CameraHelper
 from controllers.gui_controller import GuiHumanoidController
 from gui_navigation_helper import GuiNavigationHelper
-from hablab_utils import get_agent_art_obj_transform
 from sandbox_service import SandboxService
+from utils.hablab_utils import get_agent_art_obj_transform
 
 from habitat.gui.gui_input import GuiInput
 from habitat.gui.text_drawer import TextOnScreenAlignment
